Add -m (import env) flag
The -m flag allows you to import environment variables from the
environment. Because this overlaps with the behavior of -i, it
implicitly sets -i (and cannot be unset) since unsetting -i makes -m
useless.

Proved useful since I specifically wanted to be able to import PATH but
nothing else, as an example.

Change-Id: Ifaaf98cce701857e78fbe3bcd2c7d5040f940746