Allow wildcard imports from the environment
Basically, if you include * or ? in an import name, it'll be
interpreted as a wildcard. This ends up converting it to a regular
expression (gack) that should perform the same as a bash glob, but
I can't guarantee that bash wouldn't do something insane. Anyway, it's
a solution, it'd be better to replace it with something lighter weight
than a regular expression later.

Change-Id: If57c9fa47cc5d846b605086ace27678fd7717491